Extract non-DEM bake assembler
